requirements.txt
customtkinter
tkinterdnd2
faster-whisper
pyannote.audio
torch
torchaudio
huggingface_hub
numpy
matplotlib
soundfile
